MXS-3462: Propagate target updates upwards
Whenever a service is updated, all other services that use it must also
recalculate the servers they reach and the capabilities they require.

Adding a method to the mxs::Target API that exposes whether it is a server
or a service makes calculating whether a service uses another service much
simpler. An alternative way would be to do a lookup based on the name of
the mxs::Target but that would require looping over all services or
servers for each server used by a service.

The third option is to ignore the type and process all mxs::Targets but
this increases the depth of the tree by a level which in turn might
significantly increase the amount of edges that have to be traversed.

Building a graph out of the services and the other services they use is
the most efficient way to reduce the number of times a link between two
services is traversed. This comes at the cost of having to insert them
into the hashtable used to build the graph. Given that we're dealing with
pointers, the hashing cost is relatively low (one modulo operation).