MXS-3454: Do session commands inside transactions
Instead of storing them until the transaction completes, the router should
simply route them the moment they arrive. This solves the problem where
prepared statements don't work inside transactions due to the special
processing they require and which is done only when it is executed as a
session command.

In addition, this corrects the behavior to be what is documented in the
router documentation.